% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leiden_naive.R
\name{leiden_naive}
\alias{leiden_naive}
\title{Run Leiden clustering algorithm natively in R}
\usage{
leiden_naive(
  adjmat,
  resolution = 1,
  randomness = 0.01,
  n_iterations = 2L,
  partition = create_singleton_partition(nrow(adjmat))
)
}
\arguments{
\item{resolution, γ}{A parameter controlling the coarseness of the clusters}

\item{randomness, θ}{A parameter controlling the refinement of the clusters}

\item{n_iterations, n}{Number of iterations to run the Leiden algorithm. By default, 2 iterations are run.}

\item{graph, adjmat}{An adjacency matrix compatible with \code{\link[igraph]{igraph}} object or an input graph as an \code{\link[PartitionedGraph]{leiden}} object (e.g., shared nearest neighbours).}

\item{seed}{Seed for the random number generator. By default uses a random seed if nothing is specified.}
}
\value{
A partition of clusters as a vector of integers
}
\description{
Implements the Leiden clustering algorithm in R using a native R implementation. Calling reticulate to run the Python version is recommended if the python "leidenalg" and "igraph" has been installed. Returns a vector of partition indices.
}
\examples{
library("Matrix")
A <- as(matrix(0, 6, 6), "dgCMatrix")
edges <- rbind(
  c(1, 2),
  c(2, 3),
  c(3, 1),
  c(4, 5),
  c(5, 6),
  c(6, 4),
  c(1, 4)
)
apply(edges, 1, function(edge){
  u <- edge[1]
  v <- edge[2]
  A[u, v] <<- A[v, u] <<- 1
})
set.seed(1234)
partition <- leiden_naive(A, resolution = 0.25)$community
table(partition)

}
\keyword{graph}
\keyword{igraph}
\keyword{mvtnorm}
\keyword{network}
\keyword{simulation}
