% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leiden.R
\name{Leiden State Matrix}
\alias{Leiden State Matrix}
\alias{leiden}
\title{Run Leiden clustering algorithm}
\usage{
leiden(adj_mat, ...)
}
\arguments{
\item{adj_mat}{An adjacency matrix compatible with \code{\link[igraph]{igraph}} object.}

\item{...}{Parameters to pass to the Python leidenalg function.}
}
\description{
Implements the Leiden clustering algorithm in R using reticulate to run the Python version. Requires the python "leidenalg" and "igraph" modules to be installed. Returns a vector of partition indices.
}
\keyword{graph}
\keyword{igraph}
\keyword{mvtnorm}
\keyword{network}
\keyword{simulation}
